+++ lib/helpers/helpers.cpp
@@ -1,0 +1,16 @@
+#include "helpers.h"
+
+void HLT(void)
+{
+  blinkLed();
+  HLT();
+}
+void InitLed(void){
+  pinMode(LED_BUILTIN, OUTPUT);
+}
+void blinkLed(int ms = 1000){
+  digitalWrite(LED_BUILTIN, HIGH);
+  delay(ms);
+  digitalWrite(LED_BUILTIN, LOW);
+  delay(ms);
+}
