+++ lib/helpers/helpers.h
@@ -1,0 +1,5 @@
+#include <Arduino.h>
+
+void blinkLed(int ms = 1000);
+void HLT(void);
+void InitLed(void);
