+++ platformio.ini
@@ -16,3 +16,12 @@ lib_deps =
         https://github.com/adafruit/Adafruit_BME280_Library
         https://github.com/arduino-libraries/SD
         https://github.com/Naguissa/uRTCLib
+
+[env:uno]
+platform = atmelavr
+board = uno
+framework = arduino 
+lib_deps = 
+        https://github.com/adafruit/Adafruit_BME280_Library
+        https://github.com/arduino-libraries/SD
+        https://github.com/Naguissa/uRTCLib
