+++ src/drivers/SerialDriver/SerialDriver.h
@@ -1,0 +1,7 @@
+#include <Arduino.h>
+
+
+void SerialInit(void);
+void PrintFloatln(float toPrint);
+void PrintCharln(const char* toPrint);
+void PrintChar(const char* toPrint);
