+++ src/drivers/sdDriver/sdCard.h
@@ -1,0 +1,5 @@
+#include <Arduino.h>
+#include <SPI.h>
+#include <SD.h>
+#include <settings.h>
+#include <helpers.h>
