+++ src/sal/sal.h
@@ -1,17 +1,13 @@
-#include <Arduino.h>
-#include <Adafruit_BME280.h>
-#include <uRTCLib.h>
-#include <SPI.h>
-#include <SD.h>
 #include <settings.h>
+
+
+#ifdef SERIAL_DEBUG_BLE
+    #include <SoftwareSerial.h>
+#endif
 
 
 void InitLed(void); 
 void HLT(void);
-void SerialInit(void);
-void PrintFloatln(float toPrint);
-void PrintCharln(const char* toPrint);
-void PrintChar(const char* toPrint);
 void InitTime(void);
 void ReadTime(void);
 void PrintTime(void);
